+++ frontend/src/features/commons/components/TextList.vue
@@ -1,11 +1,6 @@
 <template>
   <v-list>
-    <v-list-item
-      v-for="(item, i) in items"
-      :key="i"
-      two-line
-      :href="item.link"
-    >
+    <v-list-item v-for="(item, i) in sortedItems" :key="i" two-line :href="item.link">
       <v-list-item-content>
         <v-list-item-title><b>{{ item.header }}</b></v-list-item-title>
         <v-list-item-subtitle>
@@ -20,9 +15,9 @@
 
 <script lang="ts">
 
-import {defineComponent} from "vue";
+import { defineComponent, computed } from "vue";
 
-export default defineComponent ( {
+export default defineComponent({
   name: "TextList",
   props: {
     items: {
@@ -30,15 +25,27 @@ export default defineComponent ( {
       type: Array,
       default: () => []
     }
+  },
+  setup(props) {
+    const sortedItems = computed(() => {
+      return [...props.items].sort((a, b) => {
+        return a.header.localeCompare(b.header, undefined, { sensitivity: 'base' });
+      });
+    });
+
+    return {
+      sortedItems
+    };
+
   }
 });
 </script>
 
 <style scoped>
 .item-entry {
-    word-wrap: break-word;
-    overflow-wrap: break-word;
-    white-space: pre-wrap;
-    max-width: 100%;
-  }
+  word-wrap: break-word;
+  overflow-wrap: break-word;
+  white-space: pre-wrap;
+  max-width: 100%;
+}
 </style>
